# Fix ring tokens sorting regression introduced in #3601