… pkg/ring/lifecycler.go