@bboreham 2021-02-12
After this change, are tokens sorted in disk file and in KV store?
That would be a good thing to ensure, because we could then remove sort-on-load after a couple of versions.