@pracucci 2021-02-12
> After this change, are tokens sorted in disk file and in KV store?

Yes, they should (unless I'm missing anything).

> because we could then remove sort-on-load after a couple of versions

My idea (in the PR description) was to remove it from the next version and require to upgrade to Cortex 1.7 before 1.8. What do you think? It would allow us to immediately remove the sort-on-load from `master`.