@pracucci 2021-02-12
🤞 Let's go and hope (even if hope is never a strategy)